Refatorar tela de boas-vindas para permitir ser usada em outra tela
Extraido componente de orientacao para componente global de apresentacao
