+++ lib/stores/orientacao_inicial.store.dart
@@ -5,7 +5,7 @@ import 'package:appserap/main.ioc.dart';
 import 'package:appserap/services/api.dart';
 import 'package:appserap/stores/tema.store.dart';
 import 'package:appserap/stores/usuario.store.dart';
-import 'package:appserap/ui/views/orientacao_inicial/widget_orientacao_inicial/orientacao_inicial.model.dart';
+import 'package:appserap/ui/widgets/apresentacao/apresentacao.model.widget.dart';
 import 'package:chopper/chopper.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_html/flutter_html.dart';
@@ -17,14 +17,10 @@ part 'orientacao_inicial.store.g.dart';
 class OrientacaoInicialStore = _OrientacaoInicialStoreBase with _$OrientacaoInicialStore;
 
 abstract class _OrientacaoInicialStoreBase with Store, Loggable {
-  final usuario = GetIt.I.get<UsuarioStore>();
   final _orientacaoService = GetIt.I.get<ApiService>().orientacoesIniciais;
 
   @observable
-  int pagina = 0;
-
-  @observable
-  ObservableList<OrientacaoInicialModel> listaPaginasOrientacoes = <OrientacaoInicialModel>[].asObservable();
+  ObservableList<ApresentacaoModelWidget> listaPaginasOrientacoes = <ApresentacaoModelWidget>[].asObservable();
 
   @action
   Future<void> popularListaDeOrientacoes() async {
@@ -37,32 +33,35 @@ abstract class _OrientacaoInicialStoreBase with Store, Loggable {
         body!.sort((dica1, dica2) => dica1.ordem!.compareTo(dica2.ordem!));
 
         for (var dica in body) {
-          bool mostrarHtml =
-              ((dica.imagem == null || dica.imagem!.isEmpty) && (dica.titulo == null || dica.titulo!.isEmpty));
+          bool mostrarHtml = ((dica.imagem == null || dica.imagem!.isEmpty) && (dica.titulo == null || dica.titulo!.isEmpty));
 
           if (mostrarHtml) {
-            listaPaginasOrientacoes.add(OrientacaoInicialModel(
-              titulo: '',
-              descricao: '',
-              imagem: CircularProgressIndicator(),
-              corpoPersonalizado: Html(
-                data: dica.descricao,
-                style: {
-                  '*': Style.fromTextStyle(
-                    TextStyle(
-                      fontFamily: ServiceLocator.get<TemaStore>().fonteDoTexto.nomeFonte,
-                      fontSize: ServiceLocator.get<TemaStore>().size(16),
-                    ),
-                  )
-                },
+            listaPaginasOrientacoes.add(
+              ApresentacaoModelWidget(
+                titulo: '',
+                descricao: '',
+                imagem: CircularProgressIndicator(),
+                corpoPersonalizado: Html(
+                  data: dica.descricao,
+                  style: {
+                    '*': Style.fromTextStyle(
+                      TextStyle(
+                        fontFamily: ServiceLocator.get<TemaStore>()
+                            .fonteDoTexto
+                            .nomeFonte,
+                        fontSize: ServiceLocator.get<TemaStore>().size(16),
+                      ),
+                    )
+                  },
+                ),
+                ehHTML: true,
               ),
-              ehHTML: true,
-            ));
+            );
           } else {
             String urlImagem = dica.imagem!;
 
             listaPaginasOrientacoes.add(
-              OrientacaoInicialModel(
+              ApresentacaoModelWidget(
                 titulo: dica.titulo,
                 descricao: dica.descricao,
                 imagem: Center(
@@ -87,7 +86,6 @@ abstract class _OrientacaoInicialStoreBase with Store, Loggable {
   }
 
   void dispose() {
-    pagina = 0;
-    listaPaginasOrientacoes = <OrientacaoInicialModel>[].asObservable();
+    listaPaginasOrientacoes = <ApresentacaoModelWidget>[].asObservable();
   }
 }
