--- lib/ui/views/orientacao_inicial/widget_orientacao_inicial/orientacao_inicial.model.dart
@@ -1,11 +1,0 @@
-import 'package:flutter/widgets.dart';
-
-class OrientacaoInicialModel {
-  Widget? imagem;
-  String? titulo;
-  String? descricao;
-  Widget? corpoPersonalizado;
-  bool ehHTML;
-
-  OrientacaoInicialModel({this.imagem, this.titulo, this.descricao, this.corpoPersonalizado, this.ehHTML = false});
-}
