+++ lib/ui/widgets/apresentacao/apresentacao.model.widget.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/widgets.dart';
+
+class ApresentacaoModelWidget {
+  Widget? imagem;
+  String? titulo;
+  String? descricao;
+  Widget? corpoPersonalizado;
+  bool ehHTML;
+
+  ApresentacaoModelWidget({
+    this.imagem,
+    this.titulo,
+    this.descricao,
+    this.corpoPersonalizado,
+    this.ehHTML = false,
+  });
+}
